pyproject.toml
[project]
name = "syncer"
version = "0.1.0"
description = "文件夹同步系统 - 将Linux文件夹映射到Windows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.16",
    "fastapi>=0.115.12",
    "python-socketio>=5.12.1",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
]
